remaking pipeline for trafic  simulation

+++ src/metrics.cpp
@@ -11,7 +11,7 @@
 #include <cmath>
 
 // Jain's fairness index: (Σx)² / (n · Σx²).
-double computeFairness(const std::vector<int>& hist, long /*window_packets*/, int /*channels*/) {
+double computeFairness(const std::vector<int64_t>& hist, long /*window_packets*/, int /*channels*/) {
     if (hist.empty()) {
         return 0.0;
     }
@@ -24,13 +24,13 @@ double computeFairness(const std::vector<int>& hist, long /*window_packets*/, in
 }
 
 // Normalised Pearson chi-square: ( Σ (xᵢ − e)² / e ) / window_packets ,  e = window_packets / n.
-double computeChi(const std::vector<int>& hist, long window_packets, int /*channels*/) {
+double computeChi(const std::vector<int64_t>& hist, long window_packets, int /*channels*/) {
     if (hist.empty() || window_packets == 0) {
         return 0.0;
     }
     double e = static_cast<double>(window_packets) / hist.size();
     double chi2 = 0.0;
-    for (int num : hist) {
+    for (int64_t num : hist) {
         double diff = num - e;
         chi2 += (diff * diff) / e;
     }
@@ -38,13 +38,13 @@ double computeChi(const std::vector<int>& hist, long window_packets, int /*chann
 }
 
 // Packets above the fair share, summed over overloaded channels, in per-mille of the window.
-double computeOverThreshold(const std::vector<int>& hist, long window_packets, int channels) {
+double computeOverThreshold(const std::vector<int64_t>& hist, long window_packets, int channels) {
     if (hist.empty() || window_packets == 0) {
         return 0.0;
     }
     double avg = static_cast<double>(window_packets) / channels;
     double sum = 0.0;
-    for (int num : hist) {
+    for (int64_t num : hist) {
         sum += std::max(0.0, num - avg);
     }
     return sum / (window_packets / 1000.0);
